Fixed the issue

+++ DuggaSys/fileedservice.php
@@ -51,26 +51,39 @@ if (checklogin() && $hasAccess) {
     if ($kind == 2 && isSuperUser($_SESSION['uid'] == false)) return;
 
     if (strcmp($opt, "DELFILE") === 0 && (hasAccess($userid, $cid, 'w') || isSuperUser($userid))) {
-        // Remove file link from database
-    if ($kind == 2 && isSuperUser($userid)){
-        $querystring = 'DELETE FROM fileLink WHERE fileid=:fid';
-        $query = $pdo->prepare($querystring);
-        $query->bindParam(':fid', $fid);
-        if (!$query->execute()) {
-            $error = $query->errorInfo();
-            $debug = "Error updating file list " . $error[2];
-        }
+		$counted = 0;
+		//Check if file is in use
+		$querystring0 = 'SELECT COUNT(*) counted FROM fileLink, box WHERE box.filename = fileLink.filename AND (fileLink.kind = 2 OR fileLink.kind = 3) AND fileLink.fileid=:fid ;';
+		$query0 = $pdo->prepare($querystring0);
+		$query0->bindParam(':fid', $fid);
+		if (!$query0->execute()) {
+			$error = $query0->errorInfo();
+			$debug = "Error getting file list " . $error[2];
+		}
+		$result = $query0->fetch(PDO::FETCH_OBJ);
+		$counted = $result->counted;
+		if($counted == 0){
+			// Remove file link from database
+			if ($kind == 2 && isSuperUser($userid)){
+				$querystring = 'DELETE FROM fileLink WHERE fileid=:fid';
+				$query = $pdo->prepare($querystring);
+				$query->bindParam(':fid', $fid);
+				if (!$query->execute()) {
+					$error = $query->errorInfo();
+					$debug = "Error updating file list " . $error[2];
+				}
 
-        chdir("../");
-        $currcwd = getcwd();
+				chdir("../");
+				$currcwd = getcwd();
 
-        if ($kind == 2) {
-            $currcwd .= "/courses/global/" . $filename;
-
-            if (file_exists($currcwd))
-            unlink($currcwd);
-    }
-    }
+				if ($kind == 2) {
+					$currcwd .= "/courses/global/" . $filename;
+					if (file_exists($currcwd)) unlink($currcwd);
+				}
+			}
+		}else{
+			$debug = "This file is part of a code example. Remove it from there before removing the file.";
+		}
     if($kind != 2){
         $querystring = 'DELETE FROM fileLink WHERE fileid=:fid';
         $query = $pdo->prepare($querystring);
@@ -151,19 +164,7 @@ if (checklogin() && $hasAccess) {
             }
         }
 		
-    } else if (strcmp($opt, "CHECK-IF-IN-USE") === 0) {
-		//$querystring = 'SELECT COUNT (*) FROM fileLink, box ' +
-		//+' WHERE box.filename=filelink.filename ' +
-		//+' AND filelink.kind = 2 ' +
-		//+ 'AND filelink.fileid = :fid ;';
-		$querystring = 'SELECT COUNT (*) FROM fileLink';
-        $query = $pdo->prepare($querystring);
-        $query->bindParam(':fid', $fid);
-        if (!$query->execute()) {
-            $error = $query->errorInfo();
-            $debug = "Error updating file list " . $error[2];
-        }
-	}
+    }
 	
 }
 
@@ -176,6 +177,27 @@ $files = array();
 $lfiles = array();
 $gfiles = array();
 $access = False;
+
+if (strcmp($opt, "CHECK-IF-IN-USE") === 0) {
+/*
+	$querystring = 'SELECT COUNT (*) counted FROM fileLink, box WHERE box.filename = filelink.filename AND filelink.kind = 2 AND filelink.fileid =:fid ;';
+	//$querystring = 'SELECT * FROM filelink LIMIT 2';
+    $query = $pdo->prepare($querystring);
+    $query->bindParam(':fid', getOP('fid'));
+    if (!$query->execute()) {
+        $error = $query->errorInfo();
+        $debug = "Error updating file list " . $error[2];
+    }
+	foreach ($query->fetchAll(PDO::FETCH_ASSOC) as $row) {
+		$entry = array(
+			'timesInUse' => json_encode($row['counted']),
+			//'isInUse' => $entry,
+		);
+		array_push($entries, $entry);
+	}
+	*/
+}
+
 if (checklogin() && $hasAccess) {
     $query = $pdo->prepare("SELECT fileid,filename,kind, filesize, uploaddate FROM fileLink WHERE ((cid=:cid AND vers is null) OR (cid=:cid AND vers=:vers) OR isGlobal='1') ORDER BY filename;");
     $query->bindParam(':cid', $cid);
